+++ crates/module-system/sov-modules-api/src/runtime/capabilities/auth.rs
@@ -39,12 +39,11 @@ use thiserror::Error;
 
 use crate::transaction::{
     AuthenticatedTransactionAndRawHash, Credentials, Transaction, TransactionVerificationError,
-    TransactionWithoutCall,
 };
 use crate::{
     Context, CryptoSpec, DispatchCall, ExecutionContext, FullyBakedTx, GasMeter, GasMeteringError,
-    InfallibleStateAccessor, MeteredBorshDeserialize, MeteredHasher, ProvableStateReader, RawTx,
-    Spec,
+    InfallibleStateAccessor, MeteredBorshDeserialize, MeteredBorshDeserializeError, MeteredHasher,
+    ProvableStateReader, RawTx, Spec,
 };
 
 /// The chain ID of the rollup.
@@ -297,31 +296,44 @@ pub fn authenticate<
     S: Spec,
     D: DispatchCall<Spec = S>,
 >(
-    raw_tx: &[u8],
+    mut raw_tx: &[u8],
     chain_hash: &[u8; 32],
     state: &mut Accessor,
 ) -> Result<AuthenticationOutput<S, D::Decodable, AuthorizationData<S>>, AuthenticationError> {
     let raw_tx_hash = calculate_hash::<Accessor, S>(raw_tx, state)
         .map_err(|e| AuthenticationError::OutOfGas(e.to_string()))?;
-    let (call, tx_info) = decode_sov_tx::<S, D>(raw_tx)
-        .map_err(|e| AuthenticationError::FatalError(e, raw_tx_hash))?;
-    state
-        .charge_gas(&Transaction::<D, S>::gas_cost_to_deserialize::<S>(raw_tx))
-        .map_err(|e| AuthenticationError::OutOfGas(e.to_string()))?;
 
-    let tx = tx_info.with_call(call);
+    let tx =
+        match <Transaction<D, S> as MeteredBorshDeserialize<S>>::deserialize(&mut raw_tx, state) {
+            Ok(ok) => ok,
+            Err(MeteredBorshDeserializeError::GasError(e)) => {
+                return Err(AuthenticationError::OutOfGas(format!(
+                    "Transaction deserialization run out of gas {}, tx hash {}",
+                    e, raw_tx_hash
+                )))
+            }
+            Err(MeteredBorshDeserializeError::IOError(e)) => {
+                return Err(AuthenticationError::FatalError(
+                    FatalError::DeserializationFailed(e.to_string()),
+                    raw_tx_hash,
+                ));
+            }
+        };
 
     verify_and_decode_tx::<S, D>(raw_tx_hash, tx, chain_hash, state)
 }
 
+#[cfg(feature = "native")]
 /// Decode bytes as a Sovereign SDK transaction, returning the message and tx info.
 pub fn decode_sov_tx<S: Spec, D: DispatchCall<Spec = S>>(
-    raw_tx: &[u8],
-) -> Result<(D::Decodable, TransactionWithoutCall<S>), FatalError> {
-    let tx: Transaction<D, S> =
-        borsh::from_slice(raw_tx).map_err(|e| FatalError::DeserializationFailed(e.to_string()))?;
-    let (tx, call) = tx.split();
+    mut raw_tx: &[u8],
+) -> Result<(D::Decodable, crate::transaction::TransactionWithoutCall<S>), FatalError> {
+    let tx = <Transaction<D, S> as MeteredBorshDeserialize<S>>::deserialize_without_charging_gas(
+        &mut raw_tx,
+    )
+    .map_err(|e| FatalError::DeserializationFailed(e.to_string()))?;
 
+    let (tx, call) = tx.split();
     Ok((call, tx))
 }